requirements.txt
requests
django-filter
Django==2.2.4